Fix rounding on occupancy computation.

+++ src/gov/nist/spectrumbrowser/client/DailyStatsChart.java
@@ -107,8 +107,8 @@ public class DailyStatsChart implements SpectrumBrowserCallback<String> {
 		}
 	}
 	
-	private double round(double val) {
-		return (double)(int)val*100/100;
+	private float round(double val) {
+		return (float)((int)val*100/100.0);
 	}
 
 	public void draw() {
